--- a/broker/src/main/java/io/zeebe/broker/system/configuration/backpressure/FixedCfg.java
+++ b/broker/src/main/java/io/zeebe/broker/system/configuration/backpressure/FixedCfg.java
@@ -9,7 +9,7 @@ package io.zeebe.broker.system.configuration.backpressure;
 
 import static io.zeebe.broker.system.configuration.ConfigurationUtil.checkPositive;
 
-public class FixedLimitCfg {
+public class FixedCfg {
 
   private int limit = 20;