--- a/broker/src/main/java/io/zeebe/broker/transport/backpressure/PartitionAwareRequestLimiter.java
+++ b/broker/src/main/java/io/zeebe/broker/transport/backpressure/PartitionAwareRequestLimiter.java
@@ -19,7 +19,7 @@ import com.netflix.concurrency.limits.limit.WindowedLimit;
 import io.zeebe.broker.system.configuration.backpressure.AIMDCfg;
 import io.zeebe.broker.system.configuration.backpressure.BackpressureCfg;
 import io.zeebe.broker.system.configuration.backpressure.BackpressureCfg.LimitAlgorithm;
-import io.zeebe.broker.system.configuration.backpressure.FixedLimitCfg;
+import io.zeebe.broker.system.configuration.backpressure.FixedCfg;
 import io.zeebe.broker.system.configuration.backpressure.Gradient2Cfg;
 import io.zeebe.broker.system.configuration.backpressure.GradientCfg;
 import io.zeebe.broker.system.configuration.backpressure.VegasCfg;
@@ -58,8 +58,8 @@ public final class PartitionAwareRequestLimiter {
         limit = () -> getAIMD(aimdCfg);
         break;
       case FIXED:
-        final FixedLimitCfg fixedLimitCfg = backpressureCfg.getFixedLimit();
-        limit = () -> FixedLimit.of(fixedLimitCfg.getLimit());
+        final FixedCfg fixedCfg = backpressureCfg.getFixed();
+        limit = () -> FixedLimit.of(fixedCfg.getLimit());
         break;
       case GRADIENT:
         final GradientCfg gradientCfg = backpressureCfg.getGradient();