--- a/broker/src/test/java/io/zeebe/broker/system/configuration/BackpressureCfgTest.java
+++ b/broker/src/test/java/io/zeebe/broker/system/configuration/BackpressureCfgTest.java
@@ -11,7 +11,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import io.zeebe.broker.system.configuration.backpressure.BackpressureCfg;
 import io.zeebe.broker.system.configuration.backpressure.BackpressureCfg.LimitAlgorithm;
-import io.zeebe.broker.system.configuration.backpressure.FixedLimitCfg;
+import io.zeebe.broker.system.configuration.backpressure.FixedCfg;
 import io.zeebe.broker.system.configuration.backpressure.Gradient2Cfg;
 import io.zeebe.broker.system.configuration.backpressure.GradientCfg;
 import io.zeebe.broker.system.configuration.backpressure.VegasCfg;
@@ -59,10 +59,10 @@ public final class BackpressureCfgTest {
   public void shouldSetFixedLimitCfg() {
     // when
     final BrokerCfg cfg = readConfig("backpressure-cfg");
-    final FixedLimitCfg fixedLimitCfg = cfg.getBackpressure().getFixedLimit();
+    final FixedCfg fixedCfg = cfg.getBackpressure().getFixed();
 
     // then
-    assertThat(fixedLimitCfg.getLimit()).isEqualTo(12);
+    assertThat(fixedCfg.getLimit()).isEqualTo(12);
   }
 
   @Test