… dist/src/main/config/broker.standalone.yaml.template