## Description

Fixes the mismatch between documentation and implementation of the back pressure algorithm "fixed".

Sets "fixed" as new default back pressure algorithm. Evaluated with one partition and different limits to find an appropriate value, see this comment https://github.com/zeebe-io/zeebe/issues/5113#issuecomment-682679016 .

As last evaluation I did a benchmark with three partitions and our usual load of 300 workflow instances per seconds and 12 workers.

|Name | Base | Fixed Limit |
|---------|--------|-----------------------|
|General| ![base-300wf-3p-general](https://user-images.githubusercontent.com/2758593/91726933-4948ef00-eba1-11ea-8cbd-a0e006e57c7f.png) | ![fixed-300wf-3p-general](https://user-images.githubusercontent.com/2758593/91726939-4a7a1c00-eba1-11ea-8096-72f353e4f156.png) |
| Latency | ![base-300wf-3p-latency](https://user-images.githubusercontent.com/2758593/91726935-4a7a1c00-eba1-11ea-8089-fe7a1642e96d.png) | ![fixed-300wf-3p-latency](https://user-images.githubusercontent.com/2758593/91726941-4b12b280-eba1-11ea-8283-f16248f57d43.png) |


We can see that the throughput is a bit lower then usual, but we reduced the latency by factor two. Normally the workflow instance execution time was between 2.5 - 5 seconds, with the fixed limit it is between 1 to 2.5 seconds. Furthermore the command and event processing latency is normally between 250 to 500 ms now it is between 100 to 250 ms.

<!-- Please explain the changes you made here. -->

## Related issues

<!-- Which issues are closed by this PR or are related -->

closes https://github.com/zeebe-io/zeebe/issues/5257
closes https://github.com/zeebe-io/zeebe/issues/5113

## Definition of Done

_Not all items need to be done depending on the issue and the pull request._

Code changes:
* [ ] The changes are backwards compatibility with previous versions
* [ ] If it fixes a bug then PRs are created to [backport](https://github.com/zeebe-io/zeebe/compare/stable/0.24...develop?expand=1&template=backport_template.md&title=[Backport%200.24]) the fix to the last two minor versions

Testing:
* [ ] There are unit/integration tests that verify all acceptance criterias of the issue
* [ ] New tests are written to ensure backwards compatibility with further versions
* [ ] The behavior is tested manually
* [ ] The impact of the changes is verified by a benchmark 

Documentation: 
* [ ] The documentation is updated (e.g. BPMN reference, configuration, examples, get-started guides, etc.)
* [ ] New content is added to the [release announcement](https://drive.google.com/drive/u/0/folders/1DTIeswnEEq-NggJ25rm2BsDjcCQpDape)