@deepthidevaki 2020-08-31
Thanks @Zelldon for evaluating with different limits. We thought that vegas is too conservative, so we expected that evaluating with a fixedLimit would allows us to find a higher limit value. Reducing the limit to 10 would not solve the problem encountered by first time users. Wdyt @menski @npepinpe ?