@Zelldon 2020-08-31
Ok @deepthidevaki feel free to close this if this doesn't fit. I think at least we should fix the doc and implementation mismatch for the fixedLimit algorithm.