@menski 2020-09-01
I think the limit should be higher then 10, gut feeling would be 100. I don't think it's important if the limit works well for constant throughput scenarios, but that it handles the getting started experience well, i.e. a burst or similar.