@deepthidevaki 2020-09-01
@menski,  zell has evaluated with limit 100, but the processing latency is very high then. https://github.com/zeebe-io/zeebe/issues/5113#issuecomment-682679016