@npepinpe 2020-09-01
What about AIMD as default  with a higher initial limit (e.g. 100) and target RTT (e.g. 3s)?