@menski 2020-09-01
> @menski, zell has evaluated with limit 100, but the processing latency is very high then. [#5113 (comment)](https://github.com/zeebe-io/zeebe/issues/5113#issuecomment-682679016)

Yes I think this is expected, but in this case we are pushing around 20 instances per second through the system. This is not what I expect from a getting started scenario. That's when you are reaching production use cases and have to think about the right back pressure algorithm imho. But maybe @npepinpe idea works better for getting startet and transition to production.