@deepthidevaki 2020-09-03
AIMD is not good for handling bursts, unless it is configured with a high 'minLimit'. During bursts, it aggressively reduces the limit upto minLimit (which default is 1, I think). May be a minLimit of 50 or higher might work.