@menski 2020-09-08
I split out the first part and finished it in PR #5317. I will close this PR  for now, until we decided what is the algorithm and config we want to use as default.